chore(cpp): add strict clang-tidy lint gate and clear its first findings
The C++ side had no static-analysis safety net — only compiler warnings.
This wires clang-tidy into the build as a real gate: `make tidy` lints every
solution (honoring the same `T=` substring filter as `make test`) and fails on
any finding via `--warnings-as-errors`, while the checked-in `.clang-tidy`
keeps warnings non-fatal so editors stay usable.

The check set is deliberately curated, not the firehose. A broad run drowned in
diagnostics that fight this repo's LeetCode conventions: trailing-return-type
and by-value-param checks would rewrite the verbatim problem signatures,
magic-number and bounds checks would flag every test literal and `operator[]`,
and the TEST() harness in test.hpp tripped throwing-static-init and
macro-parentheses. Those are disabled (and the header excluded) so the surviving
bugprone/modernize/performance and initializer checks carry real signal.

Turning the gate on immediately surfaced two genuine issues, fixed here so the
tree passes its own check:

  * find_missing_elements: the [min, max] span was computed as
    `max_val - min_val + 1` in `int` and only then widened to size_t — a
    misplaced widening cast that can overflow at the ±1e9 constraint bounds.
    The subtraction now happens in int64_t before narrowing.
  * maximum_length_substring: the raw `int freq[26]` frequency table becomes a
    `std::array<int, 26>`.

clang-tidy ships only with Homebrew LLVM here (not Xcode CLT), so the Makefile
auto-detects it: `command -v clang-tidy`, falling back to the brew path.

+++ cpp/src/find_missing_elements_3731.cpp
@@ -23,7 +23,7 @@ std::vector<int> findMissingElements(std::vector<int> nums) {
 // range collecting values that are still `false`.
 std::vector<int> findMissingElementsBool(std::vector<int> nums) {
     const auto [min_val, max_val] = std::ranges::minmax(nums);
-    const auto span = static_cast<std::size_t>(max_val - min_val + 1);
+    const auto span = static_cast<std::size_t>(static_cast<std::int64_t>(max_val) - min_val + 1);
 
     std::vector<bool> present(span, false);
     for (int v : nums) {
@@ -58,7 +58,7 @@ std::vector<int> findMissingElementsHashSet(const std::vector<int>& nums) {
 // range collecting values whose bit is unset.
 std::vector<int> findMissingElementsBitset(const std::vector<int>& nums) {
     const auto [min_val, max_val] = std::ranges::minmax(nums);
-    const auto span = static_cast<std::size_t>(max_val - min_val + 1);
+    const auto span = static_cast<std::size_t>(static_cast<std::int64_t>(max_val) - min_val + 1);
     const auto num_words = (span + 63) / 64;
     std::vector<std::uint64_t> bits(num_words, 0);
